services/auth_service/pyproject.toml
[tool.poetry]
name = "emma-auth-service"
version = "0.1.0"
description = "Authentication micro-service for the EMMA platform"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "auth_service" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.116.1"
uvicorn = { version = "^0.35.0", extras = ["standard"] }
psycopg2-binary = "^2.9.10"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
email-validator = "^2.1.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
httpx = "^0.28.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
